refactor(saves): extract StatusService class to saves/status/ package
StatusService owns the read-only matrix-driven save-status reporting
that powers the SAVES tab UI. Three class methods
(_get_save_status_io, _status_entry_for_local_file,
_status_entry_for_server_only) move into saves/status/__init__.py;
three pure DTO builders (_build_file_status, _status_from_action,
_resolve_chosen_server) move into saves/status/builders.py as
module-level functions.

StatusService takes a SaveService back-reference (TYPE_CHECKING
guarded) for cross-cutting transitional callbacks
(_is_save_sync_enabled, _get_server_device_id, _get_rom_save_info,
_find_save_files, _filter_server_saves_to_slot, _adopt_baseline_hash).
The matrix-iterator dependency on the soon-to-be-extracted SyncEngine
will be re-wired in Commit 7.

Public async entry points on SaveService (get_save_status,
check_save_status_background) remain as thin orchestrators; their
executor target updates to self._status._get_save_status_io.

Refs #257.
